Adds note about alternative for current Java version
This repo is still pointed at from relatively recently edited [devcenter pages](https://devcenter.heroku.com/articles/deploying-clojure-applications-with-the-heroku-leiningen-plugin). Issue https://github.com/heroku/lein-heroku/issues/5 seems to indicate the reason this plugin isn't working (for me at least) on Java 11 or 12. Based on repo activity, this plugin seems to be unmaintained.

This patch adds a note to the README that Heroku has a tool that will simply accept jars, and that a Leiningen-specific tool is probably not necessary.